+++ client/src/app/layout.tsx
@@ -3,6 +3,8 @@ import { theme } from '@/theme/override';
 import { ConfigProvider, Layout } from 'antd';
 import type { Metadata } from 'next';
 import { Navbar } from '@/components/navbar';
+import { HeaderComponent } from '@/components/header';
+import { ContentWrapper } from '@/components/content';
 
 import 'antd/dist/reset.css';
 import './globals.css';
@@ -20,7 +22,10 @@ export default function RootLayout({ children }: { children: React.ReactNode }) 
           <ConfigProvider theme={theme}>
             <Layout>
               <Navbar />
-              {children}
+              <ContentWrapper>
+                <HeaderComponent />
+                {children}
+              </ContentWrapper>
             </Layout>
           </ConfigProvider>
         </StyledComponentsRegistry>