+++ client/src/shared/constants/routes.ts
@@ -5,6 +5,13 @@ export const navbarRoutes = {
   },
 };
 
-export const otherRoutes = {};
+export const otherRoutes = {
+  article: {
+    path: '/article/:id',
+  },
+  category: {
+    path: '/category/:id',
+  },
+};
 
 export const routes = Object.assign(navbarRoutes, otherRoutes);