+++ client/src/theme/override.ts
@@ -1,6 +1,6 @@
 import type { ThemeConfig } from 'antd';
 import { Raleway, Source_Sans_3 } from 'next/font/google';
-import { colorBorder } from '.';
+import { colorBorder, colorTextSecondary } from '.';
 
 const raleway = Raleway({ subsets: ['latin', 'cyrillic-ext'] });
 const sourceSans3 = Source_Sans_3({ subsets: ['latin', 'cyrillic-ext'] });
@@ -9,7 +9,7 @@ export const theme: ThemeConfig = {
   token: {
     borderRadius: 0,
     colorPrimary: '#2ebaae',
-    colorTextSecondary: '#646464',
+    colorTextSecondary,
     colorText: '#3c3b3b',
     fontFamily: raleway.style.fontFamily,
     colorBorder,
@@ -21,9 +21,6 @@ export const theme: ThemeConfig = {
     },
     Typography: {
       fontSizeHeading2: 20,
-    },
-    Card: {
-      fontFamily: sourceSans3.style.fontFamily,
     },
     Layout: {
       colorBgHeader: '#fff',